Split out BindConfig
This commit split out the BindConfig struct into its own file to keep
mod.rs clean.
