+++ libsplinter/src/rest_api/bind_config.rs
@@ -1,0 +1,37 @@
+// Copyright 2018-2022 Cargill Incorporated
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/// Bind configuration for the REST API.
+#[derive(Clone)]
+pub enum BindConfig {
+    #[cfg(feature = "https-bind")]
+    /// A secure HTTPS binding, including certificate and key paths.
+    Https {
+        bind: String,
+        cert_path: String,
+        key_path: String,
+    },
+    /// A insecure HTTP binding.
+    Http(String),
+}
+
+impl std::fmt::Display for BindConfig {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            #[cfg(feature = "https-bind")]
+            BindConfig::Https { bind, .. } => write!(f, "{}", bind),
+            BindConfig::Http(bind) => write!(f, "{}", bind),
+        }
+    }
+}
